+++ src/main/java/shop/tryit/domain/question/dto/QuestionFormDto.java
@@ -1,21 +1,23 @@
-package shop.tryit.web.question;
+package shop.tryit.domain.question.dto;
 
 import javax.validation.constraints.NotEmpty;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
+import lombok.ToString;
 
 @Data
+@ToString
 @NoArgsConstructor
-public class QuestionSaveFormDto {
+@AllArgsConstructor
+public class QuestionFormDto {
+
+    private Long questionId;
 
     @NotEmpty(message = "제목을 입력해주세요.")
     private String title;
 
     @NotEmpty(message = "내용을 입력해주세요.")
     private String content;
-
-    @Length(min = 4, max = 4, message = "비밀번호는 4자리입니다.")
-    private String password;
 
 }
